+++ src/main/java/org/scoula/auth/dto/request/ResetPasswordRequest.java
@@ -1,4 +1,4 @@
-package org.scoula.auth.dto;
+package org.scoula.auth.dto.request;
 
 public class ResetPasswordRequest {
     private String username;
